requirements.txt
# Playwright Web Scraper - Requirements

# Core
playwright==1.40.0
python-dotenv==1.0.0

# Data handling
beautifulsoup4==4.12.0

# OpenAI (opcional - para refinamento com IA)
openai==1.3.0

# Development
pytest==7.4.3
pytest-playwright==0.4.3
black==23.12.0
